agentik-b2b-app/agents/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
loguru==0.7.2
python-dotenv==1.0.0
pydantic==2.5.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Task Queue
redis==5.0.1
aioredis==2.0.1
celery==5.3.4

# HTTP & Email
requests==2.31.0
aiofiles==23.2.1
beautifulsoup4==4.12.2
lxml==4.9.3
email-validator==2.1.0

# Templates & Reports
jinja2==3.1.2
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7

# Statistics & Analysis
numpy==1.25.2
scipy==1.11.4

# Async HTTP Client
httpx==0.25.2
aiohttp==3.9.1